Added option compression_threshold for several methods. Anyways, the best value for it seems to be 0 (what it implicitly was before.

+++ src/PlanarDiagramComplex/ArcSimplifier/ProcessArc.hpp
@@ -56,7 +56,7 @@ private:
             }
         }
          
-        if constexpr ( optimization_level < 2 )
+        if constexpr ( optimization_level < UInt8(2) )
         {
             return false;
         }
